Initial commit of the highliter code.
I decided to develop the highlighter as a stand alone tool, and then
integrate it back into qtedit4, just like qmdlib.

The code needs a lot of review, and documentation, but it works now.

The first example uses a single highliter and assigns it to a simple
text editor. This gives us a simple C++ code editor.