Fix bookmarks for hiligters without initial context
Previous commit borke the bookmarking on syntax highlighters which have
no default context. There is an assupmtion that the syntax highlighter
will add a Qate::BlockData to each QTextBlock when painting (which is
not done on those occasions).

Also added a new debug message for future debuggin, just in case.